Extract KPI evaluation logic into internal KPI service
