+++ citylearn/citylearn.py
@@ -1057,193 +1057,15 @@ class CityLearnEnv(Environment, Env):
         return self._runtime_service.parse_actions(actions)
 
     def evaluate(self, control_condition: EvaluationCondition = None, baseline_condition: EvaluationCondition = None, comfort_band: float = None) -> pd.DataFrame:
-        r"""Evaluate cost functions at current time step.
+        r"""Evaluate cost functions at current time step."""
 
-        Calculates and returns building-level and district-level cost functions normalized w.r.t. the no control scenario.
-
-        Parameters
-        ----------
-        control_condition: EvaluationCondition, default: :code:`EvaluationCondition.WITH_STORAGE_AND_PARTIAL_LOAD_AND_PV`
-            Condition for net electricity consumption, cost and emission to use in calculating cost functions for the control/flexible scenario.
-        baseline_condition: EvaluationCondition, default: :code:`EvaluationCondition.WITHOUT_STORAGE_AND_PARTIAL_LOAD_BUT_WITH_PV`
-            Condition for net electricity consumption, cost and emission to use in calculating cost functions for the baseline scenario 
-            that is used to normalize the control_condition scenario.
-        comfort_band: float, optional
-            Comfort band above dry_bulb_temperature_cooling_set_point and below dry_bulb_temperature_heating_set_point beyond 
-            which occupant is assumed to be uncomfortable. Defaults to :py:attr:`citylearn.data.EnergySimulation.DEFUALT_COMFORT_BAND`.
-        
-        Returns
-        -------
-        cost_functions: pd.DataFrame
-            Cost function summary including the following: electricity consumption, zero net energy, carbon emissions, cost,
-            discomfort (total, too cold, too hot, minimum delta, maximum delta, average delta), ramping, 1 - load factor,
-            average daily peak and average annual peak.
-
-        Notes
-        -----
-        The equation for the returned cost function values is :math:`\frac{C_{\textrm{control}}}{C_{\textrm{no control}}}` 
-        where :math:`C_{\textrm{control}}` is the value when the agent(s) control the environment and :math:`C_{\textrm{no control}}`
-        is the value when none of the storages and partial load cooling and heating devices in the environment are actively controlled.
-        """
-
-        # lambda functions to get building or district level properties w.r.t. evaluation condition
-        get_net_electricity_consumption = lambda x, c: getattr(x, f'net_electricity_consumption{c.value}')
-        get_net_electricity_consumption_cost = lambda x, c: getattr(x, f'net_electricity_consumption_cost{c.value}')
-        get_net_electricity_consumption_emission = lambda x, c: getattr(x, f'net_electricity_consumption_emission{c.value}')
-
-        # Safe division helper for KPI ratios
-        def _safe_div(control_value: float, baseline_value: float):
-            try:
-                c = control_value
-                b = baseline_value
-                # Treat None/NaN/inf as 0.0 for robust normalization on short horizons
-                def _coerce(x):
-                    try:
-                        v = float(x)
-                        return v if np.isfinite(v) else 0.0
-                    except Exception:
-                        return 0.0
-                c = _coerce(c)
-                b = _coerce(b)
-                if b == 0.0:
-                    return 1.0 if c == 0.0 else None
-                return c / b
-            except Exception:
-                return None
-
-        comfort_band = EnergySimulation.DEFUALT_COMFORT_BAND if comfort_band is None else comfort_band
-        building_level = []
-
-        for b in self.buildings:
-            if isinstance(b, DynamicsBuilding):
-                control_condition = EvaluationCondition.WITH_STORAGE_AND_PARTIAL_LOAD_AND_PV if control_condition is None else control_condition
-                baseline_condition = EvaluationCondition.WITHOUT_STORAGE_AND_PARTIAL_LOAD_BUT_WITH_PV if baseline_condition is None else baseline_condition
-
-            else:
-                control_condition = EvaluationCondition.WITH_STORAGE_AND_PV if control_condition is None else control_condition
-                baseline_condition = EvaluationCondition.WITHOUT_STORAGE_BUT_WITH_PV if baseline_condition is None else baseline_condition
-
-            discomfort_kwargs = {
-                'indoor_dry_bulb_temperature': b.indoor_dry_bulb_temperature,
-                'dry_bulb_temperature_cooling_set_point': b.indoor_dry_bulb_temperature_cooling_set_point,
-                'dry_bulb_temperature_heating_set_point': b.indoor_dry_bulb_temperature_heating_set_point,
-                'band': b.comfort_band if comfort_band is None else comfort_band,
-                'occupant_count': b.occupant_count,
-            }
-            unmet, cold, hot,\
-                cold_minimum_delta, cold_maximum_delta, cold_average_delta,\
-                    hot_minimum_delta, hot_maximum_delta, hot_average_delta =\
-                        CostFunction.discomfort(**discomfort_kwargs)
-            expected_energy = b.cooling_demand + b.heating_demand + b.dhw_demand + b.non_shiftable_load
-            served_energy = b.energy_from_cooling_device + b.energy_from_cooling_storage\
-                + b.energy_from_heating_device + b.energy_from_heating_storage\
-                    + b.energy_from_dhw_device + b.energy_from_dhw_storage\
-                        + b.energy_to_non_shiftable_load
-            ec_c = CostFunction.electricity_consumption(get_net_electricity_consumption(b, control_condition))[-1]
-            ec_b = CostFunction.electricity_consumption(get_net_electricity_consumption(b, baseline_condition))[-1]
-            zne_c = CostFunction.zero_net_energy(get_net_electricity_consumption(b, control_condition))[-1]
-            zne_b = CostFunction.zero_net_energy(get_net_electricity_consumption(b, baseline_condition))[-1]
-            ce_c = CostFunction.carbon_emissions(get_net_electricity_consumption_emission(b, control_condition))[-1]
-            ce_b = CostFunction.carbon_emissions(get_net_electricity_consumption_emission(b, baseline_condition))[-1] if sum(b.carbon_intensity.carbon_intensity) != 0 else 0
-            cost_c = CostFunction.cost(get_net_electricity_consumption_cost(b, control_condition))[-1]
-            cost_b = CostFunction.cost(get_net_electricity_consumption_cost(b, baseline_condition))[-1] if sum(b.pricing.electricity_pricing) != 0 else 0
-
-            building_level_ = pd.DataFrame([{
-                'cost_function': 'electricity_consumption_total',
-                'value': _safe_div(ec_c, ec_b),
-            }, {
-                'cost_function': 'zero_net_energy',
-                'value': _safe_div(zne_c, zne_b),
-            }, {
-                'cost_function': 'carbon_emissions_total',
-                'value': _safe_div(ce_c, ce_b),
-            }, {
-                'cost_function': 'cost_total',
-                'value': _safe_div(cost_c, cost_b),
-            }, {
-                'cost_function': 'discomfort_proportion',
-                'value': unmet[-1],
-            }, {
-                'cost_function': 'discomfort_cold_proportion',
-                'value': cold[-1],
-            }, {
-                'cost_function': 'discomfort_hot_proportion',
-                'value': hot[-1],
-            }, {
-                'cost_function': 'discomfort_cold_delta_minimum',
-                'value': cold_minimum_delta[-1],
-            }, {
-                'cost_function': 'discomfort_cold_delta_maximum',
-                'value': cold_maximum_delta[-1],
-            }, {
-                'cost_function': 'discomfort_cold_delta_average',
-                'value': cold_average_delta[-1],
-            }, {
-                'cost_function': 'discomfort_hot_delta_minimum',
-                'value': hot_minimum_delta[-1],
-            }, {
-                'cost_function': 'discomfort_hot_delta_maximum',
-                'value': hot_maximum_delta[-1],
-            }, {
-                'cost_function': 'discomfort_hot_delta_average',
-                'value': hot_average_delta[-1],
-            }, {
-                'cost_function': 'one_minus_thermal_resilience_proportion',
-                'value': CostFunction.one_minus_thermal_resilience(power_outage=b.power_outage_signal, **discomfort_kwargs)[-1],
-            }, {
-                'cost_function': 'power_outage_normalized_unserved_energy_total',
-                'value': CostFunction.normalized_unserved_energy(expected_energy, served_energy, power_outage=b.power_outage_signal)[-1]
-            }, {
-                'cost_function': 'annual_normalized_unserved_energy_total',
-                'value': CostFunction.normalized_unserved_energy(expected_energy, served_energy)[-1]
-            }])
-            building_level_['name'] = b.name
-            building_level.append(building_level_)
-
-        building_level = pd.concat(building_level, ignore_index=True)
-        building_level['level'] = 'building'
-
-        ## district level
-        # set default evaluation conditions
-        control_condition = EvaluationCondition.WITH_STORAGE_AND_PARTIAL_LOAD_AND_PV if control_condition is None else control_condition
-        baseline_condition = EvaluationCondition.WITHOUT_STORAGE_AND_PARTIAL_LOAD_BUT_WITH_PV if baseline_condition is None else baseline_condition
-
-        # District-level normalized KPIs with safe division to avoid 0/0 or div-by-zero
-        ramp_c = CostFunction.ramping(get_net_electricity_consumption(self, control_condition))[-1]
-        ramp_b = CostFunction.ramping(get_net_electricity_consumption(self, baseline_condition))[-1]
-        dlf24_c = CostFunction.one_minus_load_factor(get_net_electricity_consumption(self, control_condition), window=24)[-1]
-        dlf24_b = CostFunction.one_minus_load_factor(get_net_electricity_consumption(self, baseline_condition), window=24)[-1]
-        dlf730_c = CostFunction.one_minus_load_factor(get_net_electricity_consumption(self, control_condition), window=730)[-1]
-        dlf730_b = CostFunction.one_minus_load_factor(get_net_electricity_consumption(self, baseline_condition), window=730)[-1]
-        peak24_c = CostFunction.peak(get_net_electricity_consumption(self, control_condition), window=24)[-1]
-        peak24_b = CostFunction.peak(get_net_electricity_consumption(self, baseline_condition), window=24)[-1]
-        peak_all_c = CostFunction.peak(get_net_electricity_consumption(self, control_condition), window=self.time_steps)[-1]
-        peak_all_b = CostFunction.peak(get_net_electricity_consumption(self, baseline_condition), window=self.time_steps)[-1]
-
-        district_level = pd.DataFrame([{
-            'cost_function': 'ramping_average',
-            'value': _safe_div(ramp_c, ramp_b),
-        }, {
-            'cost_function': 'daily_one_minus_load_factor_average',
-            'value': _safe_div(dlf24_c, dlf24_b),
-        },{
-            'cost_function': 'monthly_one_minus_load_factor_average',
-            'value': _safe_div(dlf730_c, dlf730_b),
-        }, {
-            'cost_function': 'daily_peak_average',
-            'value': _safe_div(peak24_c, peak24_b),
-        }, {
-            'cost_function': 'all_time_peak_average',
-            'value': _safe_div(peak_all_c, peak_all_b),
-        }])
-
-        district_level = pd.concat([district_level, building_level], ignore_index=True, sort=False)
-        district_level = district_level.groupby(['cost_function'])[['value']].mean().reset_index()
-        district_level['name'] = 'District'
-        district_level['level'] = 'district'
-        cost_functions = pd.concat([district_level, building_level], ignore_index=True, sort=False)
-
-        return cost_functions
+        return self._kpi_service.evaluate(
+            control_condition=control_condition,
+            baseline_condition=baseline_condition,
+            comfort_band=comfort_band,
+            evaluation_condition_cls=EvaluationCondition,
+            dynamics_building_cls=DynamicsBuilding,
+        )
 
     def next_time_step(self):
         r"""Advance all buildings to next `time_step`."""
